Reordered all files and got rid of code duplication. Everything works now for the algorithm where we use complete pivot diagonals (not using bridge pivots)
